requirements-deploy.txt
aws-cdk-lib==2.62.0
constructs>=10.0.0,<11.0.0
boto3==1.20.32
botocore==1.23.32
boto3-stubs==1.20.32
requests==2.28.2